fixes issue #682 Update pull notebook by adding a new ipython notebook, the pull operations notebook.
Where we put small bites into previous notebooks, this puts
all of the pull operation examples into a single notebook with
a table of contents for the notebook.

Added a simple toc generator that operats within a single notebook
and generates a std out which can be clipped into the notebook as
a simple table of contents.  Makes notebooks larger than one page
easier to work with.

Add iter... and wbemserver documentation

Adds new notebooks for the iter... operations and the wbemserver class.

Clean up the notebooks before committing

Add new notebooks to tutorial